feat: 상세탐색 더미데이터 연결 및 레포지패턴 적용

+++ app/src/main/java/com/teamwss/websoso/ui/detailExplore/keyword/model/DetailExploreKeywordModel.kt
@@ -16,7 +16,7 @@ data class DetailExploreKeywordModel(
         val keywords: List<KeywordModel>,
     ) {
         data class KeywordModel(
-            val keywordId: Long,
+            val keywordId: Int,
             val keywordName: String,
             val isSelected: Boolean = false,
         )
